removing sorting methods out of tech review to util
