+++ src/util/salt-astronomer.js
@@ -10,4 +10,17 @@ export const getSaltAstronomerName = (username, SALTAstronomers) => {
     return name ? name.name : null;
 };
 
-export const dummyFunction = () => undefined
+/**
+ * method used for sorting Salt Astronomers by name
+ * */
+export const compareByFirstName = (a, b) => {
+  const name1 = a.name.toUpperCase();
+  const name2 = b.name.toUpperCase();
+  if (name1 < name2) {
+    return -1;
+  }
+  if (name1 > name2) {
+    return 1;
+  }
+  return 0;
+};
